+++ atlasdb-impl-shared/src/main/java/com/palantir/atlasdb/transaction/service/TransactionKVSWrapper.java
@@ -37,56 +37,10 @@ public class TransactionKVSWrapper {
         this.keyValueService = keyValueService;
     }
 
-    private static Cell getTransactionCell(long startTimestamp) {
-        return Cell.create(TransactionConstants.getValueForTimestamp(startTimestamp),
-                TransactionConstants.COMMIT_TS_COLUMN);
-    }
 
-    public Long get(Long startTimestamp) {
-        Cell cell = getTransactionCell(startTimestamp);
-        Map<Cell, Value> returnMap = keyValueService.get(TransactionConstants.TRANSACTION_TABLE,
-                                                         ImmutableMap.of(cell, MAX_TIMESTAMP));
-        if (returnMap.containsKey(cell)) {
-            return TransactionConstants.getTimestampForValue(returnMap.get(cell).getContents());
-        } else {
-            return null;
-        }
-    }
 
-    public Map<Long, Long> get(Iterable<Long> startTimestamps) {
-        Map<Long, Long> result = Maps.newHashMap();
-        Map<Cell, Long> startTsMap = Maps.newHashMap();
-        for (Long startTimestamp : startTimestamps) {
-            Cell cell = getTransactionCell(startTimestamp);
-            startTsMap.put(cell, MAX_TIMESTAMP);
-        }
 
-        Map<Cell, Value> rawResults = keyValueService.get(TransactionConstants.TRANSACTION_TABLE, startTsMap);
-        for (Map.Entry<Cell, Value> e : rawResults.entrySet()) {
-            long startTs = TransactionConstants.getTimestampForValue(e.getKey().getRowName());
-            long commitTs = TransactionConstants.getTimestampForValue(e.getValue().getContents());
-            result.put(startTs, commitTs);
-        }
 
-        return result;
-    }
 
-    // It works only if key-value store supports putUnlessExists.
-    public void putUnlessExists(long startTimestamp, long commitTimestamp) throws KeyAlreadyExistsException {
-        Cell key = getTransactionCell(startTimestamp);
-        byte[] value = TransactionConstants.getValueForTimestamp(commitTimestamp);
-        keyValueService.putUnlessExists(TransactionConstants.TRANSACTION_TABLE, ImmutableMap.of(key, value));
-    }
-
-    public void putAll(Map<Long, Long> timestampMap) throws KeyAlreadyExistsException {
-        Map<Cell, byte[]> kvMap = new HashMap<>();
-        for (Map.Entry<Long, Long> entry : timestampMap.entrySet()) {
-            kvMap.put(
-                    getTransactionCell(entry.getKey()),
-                    TransactionConstants.getValueForTimestamp(entry.getValue()));
-        }
-
-        keyValueService.put(TransactionConstants.TRANSACTION_TABLE, kvMap, 0); // This can throw unchecked exceptions
-    }
 
 }